Fixing some bugs in the tests where they were not cleaning up after themselves properly.

+++ src/java112/tests/TokenLocationSearchAnalyzerOutputTest.java
@@ -194,7 +194,16 @@ public class TokenLocationSearchAnalyzerOutputTest {
         String searchTokensFilePath = searchTokensURL.getPath();
 
         File keywordFile = new File(searchTokensFilePath);
-        keywordFile.delete();
+
+        if (keywordFile.exists()) {
+            keywordFile.delete();
+        }
+
+        File testFile = new File(testOutputFilePath);
+
+        if (testFile.exists()) {
+            testFile.delete();
+        }
 
         analyzer = null;
     }
